عمل تعديل على التقارير في الادمن لمبيعات المصانع وتقرير للوحة تحكم المصنع بعرض اجمال املبغ المكتسب من الاحالة
عمل تعديل على التقارير في الادمن لمبيعات المصانع وتقرير للوحة تحكم المصنع بعرض اجمالي المبلغ المكتسب من الاحالة

+++ database/migrations/2025_01_25_125429_add_languge_in_table_banners.php
@@ -25,10 +25,13 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('banners', function (Blueprint $table) {
-            //
-            $table->dropColumn('language');
+        if(!Schema::hasColumn("banners","language")){
 
-        });
+            Schema::table('banners', function (Blueprint $table) {
+                //
+                $table->dropColumn('language');
+
+            });
+        }
     }
 };
